Disable dynamic linking for Cabal
I've switched to using Cabal installed via Nix. With dynamic linking
enabled, this resulted in errors like the following, when trying to run
`haskell-language-server`:

``` Cannot load -dynamic objects when GHC is built the normal way

To fix this, either:

(1) Use -fexternal-interpreter, or

(2) Build the program twice: once the normal way, and then with -dynamic
using -osuf to set a different object file suffix.
```